recovery: do not keep xlog open if eof reached
If we reached EOF while reading an xlog file on recovery, we don't need
to keep it open, as we won't read anything from it. However, we don't
close such xlog to avoid rescanning it over and over again until a new
xlog file created. This is ugly, besides it doesn't allow us to invoke
garbage collection right after box.snapshot() w/o getting "file deleted
under our feet" warning. Instead of keeping the cursor open, let's add
eof flag to the recovery struct - if the flag is set, the last xlog file
was read to the end and so the next xlog file should have xlog equal to
recovery->vclock.