vinyl: zap include_deleted flag argument of vy_recovery_iterate
This flag is needed to skip records corresponding to dropped and
incomplete runs on recovery, because given VY_LOG_CREATE_RUN the
vy_recovery_iterate() callback can't tell if the run is used or
going to be dropped. This flag needlessly complicates the code of
vy_recovery_iterate(), which is supposed to simply replay all records
recovered from the log, not try to be smart. Let's get rid of it and
instead add a hint to vy_log_record indicating that a run created by
VY_LOG_CREATE_RUN is going to be dropped.